Do not clear states for multi-widget checkbox
A checkbox group can have one name and multiple widgets that are valid to
be set independently on a form. However, when setting such a field passing
different valid values only the last one stays on visually, while the
previous ones get erased.

This fix removes such clearing. To actually clear a particular widget some
other mechanism is required, for example, passing a value to use and
additional flag to indicate it needs to be cleared, rather than set.