SUP-1395: Add 'SuppressTagging' property to PdfContentByte.
- Declare a boolean property 'SuppressTagging' on PdfContentByte.
- Modify the bool 'IsTagged()' method to return true only if SuppressTagging is false.

Background:
By default, when new content is added to a PdfContentByte, iText wraps it in a Begin/End marked content sequence,
and inserts a reference to the new MCID at the end of the page's tagged structure tree.
An application might prefer to disable this behavior and manage the document structure itself.
This property allows an application to prevent iText from adding any marked content or manipulating the structure tree.