Port [6d00a4a069] Avoid class cast exception because of the possible PdfIndirectReference in place of the normal appearance pdf object
Do not check if BBox and Rect are different for the field and it's appearance
stream if appearance stream on fields flattenning is available only through
PdfIndirectReference. In this case we assume that appearance was written by
iText and therefore it should be OK. This happens for instance after initial
creation of the field appearance in the same run with flattening.
These changes are connected with commit:
[f9b3338a60] "Fix flattening form fields in case of normal appearance contains PdfIndirectReference. Suggested by https://github.com/lbonco"

DEV-1927