Commit HF - Adition of WebProxy
Added WebProxy to TSAClientBouncyCastle.
It is required when working behind a proxy server.

Some how the java version has implemented the proxy but not the .net version.
Also i had a look to the IText 7 Core (at least the on the NuGet), also did not have the implementation of proxy in this class.

The proxy will be used when on the webrequest in  GetTSAResponse() method.